requirements.txt
amqp==2.2.2
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.6
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap3-datetimepicker-2==2.5.0
django-crispy-forms==1.7.0
django-datetime-widget==0.9.3
django-sslserver==0.20
djangorestframework==3.7.0
future==0.16.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
kombu==4.1.0
oauthlib==2.0.4
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
